refactor(frontend): make useUsage internal to UsageInfo, accept eventBus and className props
UsageInfo container now accepts eventBus (manages usage state internally)
and className (forwarded to view). Consumers no longer need useUsage hook
directly. Returns null when no usage data available.
