+++ apps/frontend/src/pages/chat/components/StreamingMessageDisplay/components/MessageList/components/SubagentDisclosure/styles.module.css
@@ -127,7 +127,7 @@
   font-weight: 500;
 }
 
-.usageWrapper {
+.usage {
   font-size: 0.75rem;
   color: var(--muted);
   padding: 6px 4px 0;
